+++ l10n_br_sale/models/__init__.py
@@ -2,6 +2,6 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from . import res_company
-from . import res_config
+from . import res_config_settings
 from . import sale
 from . import sale_line
