+++ l10n_br_sale/models/res_config_settings.py
@@ -1,14 +1,30 @@
 # Copyright (C) 2014  Renato Lima - Akretion
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from odoo import fields, models
+from odoo import api, fields, models
 
 
-class ResConfigSettings(models.TransientModel):
-    _inherit = 'res.config.settings'
+class SaleConfiguration(models.TransientModel):
+    _inherit = "sale.config.settings"
+
+    def _get_default_copy_note(self):
+        ir_values = self.env["ir.values"]
+        comp_id = self.env.user.company_id.id
+        return ir_values.get_default("sale.order", "copy_note", company_id=comp_id)
 
     copy_note = fields.Boolean(
-        string='Copy notes in Fiscal documents',
-        related='company_id.copy_note',
-        readonly=False,
+        u"Copiar Observações nos Documentos Fiscais", default=_get_default_copy_note
     )
+
+    @api.multi
+    def set_sale_defaults(self):
+        return (
+            self.env["ir.values"]
+            .sudo()
+            .set_default(
+                "sale.order",
+                "copy_note",
+                self.copy_note,
+                company_id=self.env.user.company_id.id,
+            )
+        )
