+++ l10n_br_sale/views/res_config_settings_view.xml
@@ -1,24 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <record id="l10n_br_sale_res_config_settings_form" model="ir.ui.view">
-        <field name="name">l10n_br_sale.res.config.settings.form</field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="sale.res_config_settings_view_form"/>
+    <record id="l10n_br_sale_res_config_form" model="ir.ui.view">
+        <field name="name">l10n_br_sale.res.config.form</field>
+        <field name="model">sale.config.settings</field>
+        <field name="inherit_id" ref="sale.view_sales_config"/>
         <field name="arch" type="xml">
-            <xpath expr="//div[@id='sales_settings_invoicing_policy']" position="after">
-                <div class="col-xs-12 col-md-6 o_setting_box">
-                    <div class="o_setting_left_pane">
-                        <field name="copy_note"/>
-                    </div>
-                    <div class="o_setting_right_pane">
-                        <label for="copy_note"/>
-                        <div class="text-muted">
-                            Copy sale order notes in fiscal document
+            <group name="customers" position="after">
+                <group>
+                    <label for="id" string="Observação"/>
+                    <div>
+                        <div name="copy_note">
+                            <field name="copy_note" class="oe_inline"/>
+                            <label for="copy_note"/>
                         </div>
                     </div>
-                </div>
-            </xpath>
+                </group>
+            </group>
         </field>
     </record>
