--- l10n_br_sale/views/res_config_view.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="l10n_br_sale_res_config_form" model="ir.ui.view">
-        <field name="name">l10n_br_sale.res.config.form</field>
-        <field name="model">sale.config.settings</field>
-        <field name="inherit_id" ref="sale.view_sales_config"/>
-        <field name="arch" type="xml">
-            <group name="customers" position="after">
-                <group>
-                    <label for="id" string="Observação"/>
-                    <div>
-                        <div name="copy_note">
-                            <field name="copy_note" class="oe_inline"/>
-                            <label for="copy_note"/>
-                        </div>
-                    </div>
-                </group>
-            </group>
-        </field>
-    </record>
-
-</odoo>
